move lib out of host implementation
